feat: added migration for new column for crop name
